Use SDK exclusively for net I/O, move to docker

+++ make.sh
@@ -1,17 +1,17 @@
 #!/usr/bin/env bash
 set -euo pipefail
-unset CDPATH; cd "$( dirname "${BASH_SOURCE[0]}" )"; cd "`pwd -P`"
+unset CDPATH; cd "$( dirname "${BASH_SOURCE[0]}" )"; cd "$(pwd -P)"
 
 # Project settings: package name, test packages (if different), Go & Glide versions, and cross-compilation targets
 pkg="flywheel.io/fw"
-testPkg=""
-coverPkg=""
-goV=${GO_VERSION:-"1.8"}
+testPkg="flywheel.io/fw"
+coverPkg="flywheel.io/fw"
+goV=${GO_VERSION:-"1.8.1"}
 minGlideV="0.12.3"
-targets=( "darwin/amd64" "linux/amd64" "windows/amd64" )
+targets=( "linux/amd64" "darwin/amd64" "windows/amd64" )
 #
 
-fatal() { echo -e $1; exit 1; }
+fatal() { echo -e "$1"; exit 1; }
 
 # Check that this project is in a gopath
 test -d ../../../src || fatal "This project must be located in a gopath.\nTry cloning instead to \"src/$pkg\"."
@@ -23,11 +23,12 @@ localOs=$( uname -s | tr '[:upper:]' '[:lower:]' )
 # Load GNU coreutils on OSX
 if [[ "$localOs" == "darwin" ]]; then
 	# Check requirements: g-prefixed commands are available if brew packages are installed.
-	which brew gsort gsed gfind > /dev/null || fatal "On OSX, homebrew is required. Install from http://brew.sh\nThen, run 'brew install bash coreutils findutils gnu-sed' to install the necessary tools."
+	hash brew gsort gsed gfind 2>/dev/null || fatal "On OSX, homebrew is required. Install from http://brew.sh\nThen, run 'brew install bash coreutils findutils gnu-sed' to install the necessary tools."
 
 	# Load GNU coreutils, findutils, and sed into path
 	suffix="libexec/gnubin"
-	export PATH="$(brew --prefix coreutils)/$suffix:$(brew --prefix findutils)/$suffix:$(brew --prefix gnu-sed)/$suffix:$PATH"
+	export PATH
+	PATH="$(brew --prefix coreutils)/$suffix:$(brew --prefix findutils)/$suffix:$(brew --prefix gnu-sed)/$suffix:$PATH"
 
 	# OSX has shasum. CentOS has sha1sum. Ubuntu has both.
 	alias sha1sum="shasum -a 1"
@@ -47,7 +48,7 @@ prepareGo() {
 	filterError='s/'"I don't have any idea what to do with"'/Download or install failed for go/g;'
 
 	# Install go, clearing tempdir before & after, with nice messaging.
-	test -f $src || (
+	test -f "$src" || (
 		echo "Downloading go $goV..."
 		rm -rf $GIMME_TMP; mkdir -p $GIMME_TMP
 
@@ -59,7 +60,7 @@ prepareGo() {
 	)
 
 	# Load installed go and prepare for compiled tools
-	source $src
+	source "$src"
 	export PATH=$GOPATH/bin:$PATH
 }
 
@@ -76,8 +77,8 @@ prepareGlide() {
 
 	installGlide() {
 		echo "Downloading glide $minGlideV or higher..."
-		mkdir -p $GOPATH/bin
-		rm -f $GOPATH/bin/glide
+		mkdir -p "$GOPATH/bin"
+		rm -f "$GOPATH/bin/glide"
 		curl -sL https://glide.sh/get | bash
 	}
 
@@ -92,7 +93,7 @@ prepareGlide() {
 		generateGlideHash > $glideHashFile
 	}
 
-	test -x $GOPATH/bin/glide || installGlide
+	test -x "$GOPATH/bin/glide" || installGlide
 
 	# Check the current glide version against the minimum project version
 	currentVersion=$(glide --version | cut -f 3 -d ' ' | tr -d 'v')
@@ -105,15 +106,21 @@ prepareGlide() {
 
 	# If glide components are missing, or cache is out of date, run
 	test -f glide.lock -a -d vendor -a -f $glideHashFile || runGlideInstall
-	test `cat $glideHashFile` = `generateGlideHash` || runGlideInstall
+	test "$(cat $glideHashFile)" = "$(generateGlideHash)" || runGlideInstall
 }
 
 build() {
-	extraLdFlags=${1:-}
+	package=${1:-$pkg}
+	extraLdFlags=${2:-}
+
+	# Check glide state, unless this is called from cross-build, which does so once already.
+	if [ -z "$extraLdFlags" ]; then
+		prepareGlide
+	fi
 
 	# Clean out the absolute-path pollution, and highlight source code filenames.
 	filterAbsolutePaths="s#$GOROOT/src/##g; s#$GOPATH/src/##g; s#$pkg/vendor/##g; s#$PWD/##g;"
-	highlightGoFiles="s#[a-zA-Z0-9_]*\.go#$(tput setaf 1)&$(tput sgr0)#;"
+	highlightGoFiles="s#[a-zA-Z0-9_]*\.go#$(tty -s && tput setaf 1 2>/dev/null || true)&$(tty -s && tput sgr0 2>/dev/null || true)#;"
 
 	# Go install uses $GOPATH/pkg to cache built files. Faster than go build.
 	#
@@ -124,18 +131,41 @@ build() {
 	# The only alternative is *recompiling the standard library N times every build*.
 	# Gimmie, provisioned above, provides a safely writable Go install in your homedir.
 	# https://dave.cheney.net/2015/08/22/cross-compilation-with-go-1-5
-	go install -v -ldflags "-s $extraLdFlags" $pkg 2>&1 | sed "$filterAbsolutePaths $highlightGoFiles"
+	go install -v -ldflags "-s $extraLdFlags" "$package" 2>&1 | sed "$filterAbsolutePaths $highlightGoFiles"
 }
 
-crossBuild() {
+cross() {
+	package=${1:-$pkg}
+	prepareGlide
+
 	# Placed here, instead of at the top, since it's the only place we need it
 	localArch=$( uname -m | sed 's/x86_//; s/i[3-6]86/32/' )
 
 	# For release builds, detect useful information about the build.
 	# Fails safely & silently. Declare & use these strings in your main!
-	BuildHash=$( which git  > /dev/null && git rev-parse --short HEAD 2> /dev/null || echo "unknown" )
-	BuildDate=$( which date > /dev/null && date "+%Y-%m-%d %H:%M"     2> /dev/null || echo "unknown" )
+	BuildHash=$( hash git  2> /dev/null && git rev-parse --short HEAD 2>/dev/null || echo "unknown" )
+	BuildDate=$( hash date 2> /dev/null && date "+%Y-%m-%d %H:%M"     2>/dev/null || echo "unknown" )
 	# Datestamp is ISO 8601-ish, without seconds or timezones.
+
+	# Versions of UPX prior to 3.92 had compatibility issues with OSX 10.12 Sierra.
+	# Don't compress the binaries unless UPX is present and new enough.
+	#
+	# https://upx.github.io/upx-news.txt
+	# https://apple.stackexchange.com/questions/251808/this-upx-compressed-binary-contains-an-invalid-mach-o-header-and-cannot-be-load
+	minUPXv="3.92"
+	useUPX=false
+
+	# Check UPX version
+	if hash upx 2>/dev/null; then
+		currentUPXv=$(upx --version | head -n 1 | cut -f 2 -d ' ')
+		floorVersion=$(echo -e "$minUPXv\n$currentUPXv" | sort -V | head -n 1)
+
+		if [[ $minUPXv == $floorVersion ]]; then
+			useUPX=true
+		else
+			echo "Warning: your UPX version is too old and cannot compress OSX binaries correctly. Disabling."
+		fi
+	fi
 
 	for target in "${targets[@]}"; do
 
@@ -144,14 +174,10 @@ crossBuild() {
 		os=${targetSplit[0]}; arch=${targetSplit[1]}
 
 		echo -e "\n-- Building $os $arch --"
-		GOOS=$os GOARCH=$arch build "-X main.BuildHash=$BuildHash -X 'main.BuildDate=$BuildDate'"
+		GOOS=$os GOARCH=$arch build "$package" "-X main.BuildHash=$BuildHash -X 'main.BuildDate=$BuildDate'"
 
-		# Versions of UPX prior to 3.92 had compatibility issues with OSX 10.12 Sierra.
-		# Could augment this with a UPX version check, and compress if local UPX is modern enough.
-		#
-		# https://upx.github.io/upx-news.txt
-		# https://apple.stackexchange.com/questions/251808/this-upx-compressed-binary-contains-an-invalid-mach-o-header-and-cannot-be-load
-		if [[ "$os" != "darwin" ]]; then
+
+		if $useUPX; then
 			if [[ "$localOs" == "$os" && "$arch" =~ .*$localArch ]] ; then
 				path="$GOPATH/bin/"
 			else
@@ -159,64 +185,94 @@ crossBuild() {
 			fi
 
 			binary=$( find "$path" -maxdepth 1 | grep -E "${pkg##*/}(\.exe)*" | head -n 1 )
-
-			which upx > /dev/null && nice upx -q $binary 2>&1 | grep -- "->" || true
+			nice upx -q "$binary" 2>&1 | grep -- "->" || true
 		fi
 
 		# If this system is the current build target, copy the binary to a build folder.
 		# Makes it easier to export a cross-build.
-
 		if [[ "$localOs" == "$os" && "$arch" =~ .*$localArch ]] ; then
 			path="$GOPATH/bin/"
 			binary=$( find "$path" -maxdepth 1 | grep -E "${pkg##*/}(\.exe)*" | head -n 1 )
 
 			mkdir -p "$GOPATH/bin/${os}_${arch}/"
-			cp $binary "$GOPATH/bin/${os}_${arch}/"
+			cp "$binary" "$GOPATH/bin/${os}_${arch}/"
 		fi
 	done
 
-	which upx > /dev/null || ( echo "UPX is not installed; did not compress binaries." )
+	hash upx 2>/dev/null || ( echo "UPX is not installed; did not compress binaries." )
+}
+
+prepareCrossBuild() {
+	prepareGo
+
+	flag="$GOROOT/.custom-flags/stdlib-cross-compiled"
+	test -f "$flag" && return 0
+
+	# Filter out packages that are esoteric, require linking, unexported, or outside the stdlib
+	template='{{.ImportPath}}${{.Standard}}'
+	filter='^(cmd|crypto/x509|debug|go/(build|types)|plugin|runtime|syscall|testing)|internal|vendor|^(net|os/user)$|false$'
+	packages=( $( go list -f $template ... | grep -v -E $filter | cut -f 1 -d '$' ) )
+
+	# Create a dummy Go file that imports every package, so we can cross-compile them ¯\_(ツ)_/¯
+	# This is mainly useful if you'd like to generate these object files, and then cache them for later.
+	# Results in faster cross-compile for your CI build!
+	folder="crossCompileStdlib"; mkdir -p "$folder"
+	tempfile="$folder/main.go"
+	echo -e "package main\nimport (" > $tempfile
+	for package in "${packages[@]}"; do echo "	_ \"$package\"" >> $tempfile; done
+	echo -e ")\nfunc main() { }" >> $tempfile
+
+	cross "$pkg/crossCompileStdlib"
+	rm -rf $folder
+	find "$GOPATH/bin" -executable -type f | grep "crossCompileStdlib" | xargs rm
+	mkdir -p "$(dirname "$flag")"; touch "$flag"
 }
 
 # Some go tools take package names. Some take file names. Some like the pacakge prefix. Some don't.
 # None of them omit the vendor directory like they should.
 # What follows is a result of these problems.
 # https://github.com/golang/go/issues/19090
+#
+# Check for resolution:
+# git clone git@github.com:golang/go.git; cd go; git branch --contains fa1d54c2edad607866445577fe4949fbe55166e1
 
 
 # List packages with source code: packageA, packageB... First argument is a prefix (optional)
 listPackages() {
 	prefix=${1:-}
-	find . -name '*.go' -printf "%h\n" | sed '/^\.\/vendor/d; s#^\./##g; /\./d; s#^#'$prefix'#g' | sort --unique
+	find . -name '*.go' -printf "%h\n" | sed '/^\.\/vendor/d; s#^\./##g; /\./d; s#^#'"$prefix"'#g' | sort --unique
 }
 
 # List files in the main package. First argument is a prefix (optional)
 listBaseFiles() {
 	prefix=${1:-}
-	find -maxdepth 1 -type f -name '*.go' | sed 's#^\./##g; s#^#'$prefix'#g'
+	find -maxdepth 1 -type f -name '*.go' | sed 's#^\./##g; s#^#'"$prefix"'#g'
 }
 
 format() {
-	gofmt -w -s $(listPackages; listBaseFiles)
+	prepareGo
+	gofmt -w -s $(listPackages && listBaseFiles)
 }
 
 formatCheck() {
-	badFiles=(`gofmt -l -s $(listPackages; listBaseFiles)`)
+	prepareGo
+	badFiles=($(gofmt -l -s $(listPackages | grep -v bridge/dist && listBaseFiles)))
 
 	if [[ ${#badFiles[@]} -gt 0 ]]; then
-		echo "The following files need formatting: " ${badFiles[@]}
+		echo "The following files need formatting: ${badFiles[*]}"
 		exit 1
 	fi
 }
 
 _test() {
+	prepareGlide
 	hideEmptyTests="/\[no test files\]$/d; /^warning\: no packages being tested depend on /d; /^=== RUN   /d;"
 
 	# If testing a single package, coverprofile is availible.
 	# Set which package to test and which package to count coverage against.
 
 	if [[ $testPkg == "" ]]; then
-		go test -v -cover "$@" $pkg $(listPackages $pkg/) 2>&1 | sed -r "$hideEmptyTests"
+		go test -v -cover "$@" "$pkg" $(listPackages $pkg/) 2>&1 | sed -r "$hideEmptyTests"
 	else
 		go test -v -cover -coverprofile=.coverage.out -coverpkg $coverPkg "$@" $testPkg 2>&1 | sed -r "$hideEmptyTests"
 
@@ -224,57 +280,40 @@ _test() {
 	fi
 }
 
+clean() {
+	# Remove all build state
+	prepareGo
+	rm -rvf "$GOPATH/pkg" "${GOPATH:?}/bin/${pkg##*/}"
+}
 
-cmd=${1:-"build"}
-shift || true
+showEnv() {
+	prepareGlide 1>&2
+	(go env; echo "PATH=$PATH") | sed 's/^/export /g'
+}
 
+cmd=${1:-"build"}; shift || true
 case "$cmd" in
-	"go" | "godoc" | "gofmt") # Go commands in project context
-		prepareGo
-		$cmd "$@"
-		;;
-	"glide") # Glide commands in project context
-		prepareGlide
-		glide "$@"
-		cleanGlideLockfile
-		;;
-	"make") # Just build
-		prepareGlide
-		build
-		;;
-	"build") # Full build (the default)
-		prepareGlide
-		build
-		format
-		;;
-	"format") # Format code
-		prepareGo
-		format
-		;;
-	"clean") # Remove build state
-		prepareGo
-		rm -rvf $GOPATH/pkg $GOPATH/bin/${pkg##*/}
-		;;
-	"test") # Run tests
-		prepareGlide
-		_test "$@"
-		;;
+	"go" | "godoc" | "gofmt")
+		prepareGo; $cmd "$@";;
+
+	"glide")
+		prepareGlide; glide "$@"; cleanGlideLockfile;;
+
+	"test")
+		_test "$@";;
+
 	"env") # Load environment!   eval $(./make.sh env)
-		prepareGlide 1>&2
-		(go env; echo "PATH=$PATH") | sed 's/^/export /g'
-		;;
-	"ci") # Monolithic CI target
-		prepareGlide
-		build
-		formatCheck
-		_test -race
-		;;
-	"cross") # Cross-compile to every platform
-		prepareGlide
-		crossBuild
-		;;
+		showEnv;;
+
+	"goserve") # Run godoc in server mode and open the docs
+		prepareGo
+		echo "Serving documentation on http://localhost:6060 ..."
+		( sleep 0.5; hash firefox 2> /dev/null && firefox http://localhost:6060/pkg/$pkg ) &
+		godoc -http :6060;;
+
 	*)
-		echo "Usage: ./make.sh {go|godoc|gofmt|glide|make|build|format|clean|test|env|ci|cross}"
-		exit 1
-		;;
+		type "${cmd}" >/dev/null 2>&1 && eval "${cmd}" || (
+			echo "Usage: ./make.sh {go|godoc|gofmt|glide|build|format|clean|test|env|ci|cross}"
+			exit 1
+		);;
 esac
